fixed some issue in compilation. Line still needs to be fixed
