--- frontend/src/app/services/data/data.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { DataService } from './data.service';
-
-describe('DataService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: DataService = TestBed.get(DataService);
-    expect(service).toBeTruthy();
-  });
-});
