+++ frontend/src/app/services/mock-childs.ts
@@ -1,8 +1,3 @@
 import { Child } from '../models/child'
 
-export const CHILDS: Child[] = [
-    { name: "Mariuzza", present: false },
-    { name: "Giacomino", present: false },
-    { name: "Antonello", present: false },
-    { name: "Paolo", present: false },
-]
+export const CHILDS: Child[] = []
